fix: PHP 8.4 compatibility for core and PHPUnit tests
Harden string/numeric type handling across wp-includes and wp-admin for
PHP 8.4 strict type requirements. Update PHPUnit tests and fixtures for
Imagick API changes, deprecation expectations, meta ID validation, and
filesystem teardown.

+++ src/wp-includes/PHPMailer/PHPMailer.php
@@ -3843,7 +3843,7 @@ class PHPMailer
         for ($i = 0; $i < $mb_length; $i += $offset) {
             $lookBack = 0;
             do {
-                $offset = $avgLength - $lookBack;
+                $offset = (int) ($avgLength - $lookBack);
                 $chunk = mb_substr($str, $i, $offset, $this->CharSet);
                 $chunk = base64_encode($chunk);
                 ++$lookBack;
